Jcarlos/general improvements

+++ src/pages/CarriersPage.tsx
@@ -35,41 +35,40 @@ export const CarriersPage = () => {
   const [isOpenModalInfo, setIsOpenModalInfo] = useState<boolean>(false);
   const [titleModalInfo, setTitleModalInfo] = useState<string>("");
 
+  const [isLoadingTable, setIsLoadingTable] = useState<boolean>(false);
   const [isLoading, setIsLoading] = useState<boolean>(false);
-  const [action, setAction] = useState<boolean>(false);
   const [errorMessage, setErrorMessage] = useState<string>();
 
   const toggleModal = (value: boolean) => setIsOpenModal(value);
   const toggleModalInfo = (value: boolean) => setIsOpenModalInfo(value);
 
   const getAllCarriers = async () => {
-    setIsLoading(true);
+    setIsLoadingTable(true);
     try {
       const res = await getAllData("carriers");
       const data: DataRowCarriers[] = res.data!;
-      if (!data) setCarriersData([]);
       setCarriersData(data);
     } catch (error) {
-      console.log(error);
+      setCarriersData([]);
+    } finally {
+      setIsLoadingTable(false);
     }
-    setIsLoading(false);
   };
   const getClientsForCarrier = async () => {
     try {
       const res = await getAllData("clients/approved-without-carrier");
       const data: DataRowCarriers[] = res.data!;
-      if (!data) setClientsForCarrier([]);
       setClientsForCarrier(data);
     } catch (error) {
-      const { message } = error as ApiResponse;
-      if (message === "No se encontró ningún cliente que pueda ser portador")
-        setClientsForCarrier([]);
+      setClientsForCarrier([]);
     }
   };
   useEffect(() => {
     getAllCarriers();
+  }, []);
+  useEffect(() => {
     getClientsForCarrier();
-  }, [action]);
+  }, [carriersData]);
 
   const handleCreate = async (data: ICarrierForm) => {
     setIsLoading(true);
@@ -78,19 +77,15 @@ export const CarriersPage = () => {
         ...data,
         placement_time: formatTime12to24(data.placement_time),
       });
-
-      if (res.success) {
-        toggleModal(false);
-        setAction(!action);
-        alertTimer(`El portador se ha agregado`, "success");
-        setErrorMessage("");
-      }
+      toggleModal(false);
+      setCarriersData((prev) => [...prev, res.data!]);
+      alertTimer(`El portador se ha agregado`, "success");
+      setErrorMessage("");
     } catch (error) {
-      const err = error as ApiResponse;
-      if (err) setErrorMessage(err.message!);
-      alertTimer(`Ha ocurrido un error.`, "error");
+      handleError(error as ApiResponse);
+    } finally {
+      setIsLoading(false);
     }
-    setIsLoading(false);
   };
   const handleUpdate = async (data: ICarrierForm) => {
     setIsLoading(true);
@@ -99,35 +94,48 @@ export const CarriersPage = () => {
         ...data,
         placement_time: formatTime12to24(data.placement_time),
       });
-
       if (res.success) {
+        const updatedCarrierData: DataRowCarriers = res.data!;
+        setCarriersData((prev) =>
+          prev.map((carrier) =>
+            carrier.id === carrierID
+              ? { ...carrier, ...updatedCarrierData }
+              : carrier
+          )
+        );
         toggleModal(false);
-        setAction(!action);
         alertTimer(`El portador se ha actualizado`, "success");
         setErrorMessage("");
       }
     } catch (error) {
-      const err = error as ApiResponse;
-      if (err) setErrorMessage(err.message!);
-      alertTimer(`Ha ocurrido un error.`, "error");
+      handleError(error as ApiResponse);
+    } finally {
+      setIsLoading(false);
     }
-    setIsLoading(false);
   };
 
   const handleDelete = (id: number) => {
     const confirm = confirmChange({
       title: "Eliminar Portador",
-      text: `¿Está seguro de querer eliminar el Portador con el ID ${id}?. Este cambio es irreversible.`,
+      text: `¿Está seguro de querer eliminar el Portador con el ID ${id}?. Esto también eliminará la operación que corresponde a este Portador.`,
       confirmButtonText: "Eliminar",
       confirmButtonColor: "red",
     });
     confirm.then(async (res) => {
       if (res.success) {
         try {
+          const deleteOp = await deleteData("operations", id);
+          if (!deleteOp.success) {
+            alertTimer("No se pudo eliminar la operación", "error");
+            return;
+          }
           const response = await deleteData("carriers", id);
-          if (response.success)
+          if (response.success) {
+            setCarriersData((prev) =>
+              prev.filter((carrier) => carrier.id !== id)
+            );
             alertTimer("El portador ha sido eliminado", "success");
-          setAction(!action);
+          }
         } catch (error) {
           const err = error as ApiResponse;
           alertTimer(err.message, "error");
@@ -135,16 +143,22 @@ export const CarriersPage = () => {
       }
     });
   };
+  const handleError = (error: ApiResponse) => {
+    if (error) setErrorMessage(error.message!);
+    alertTimer("Ha ocurrido un error", "error");
+  };
 
   const columns: TableColumn<DataRowCarriers>[] = [
     {
       name: "Nombre",
       selector: (row) => row.name,
       sortable: true,
+      wrap: true,
     },
     {
       name: "Zona de Residencia",
       selector: (row) => row.residence_area,
+      wrap: true,
     },
     {
       name: "Brazalete Electrónico",
@@ -153,14 +167,17 @@ export const CarriersPage = () => {
     {
       name: "BEACON",
       selector: (row) => row.beacon,
+      width: "120px",
+      wrap: true,
     },
     {
       name: "Cargador Inalámbrico",
       selector: (row) => row.wireless_charger,
     },
     {
-      name: "Nombre del Instalador",
+      name: "Nom. Instalador",
       selector: (row) => row.installer_name,
+      wrap: true,
     },
     {
       name: "Acciones",
@@ -203,9 +220,10 @@ export const CarriersPage = () => {
           }
           toggleModal(value);
           setCarrierData(null);
+          setCarrierID(null);
           setTitleModal("Agregar Portador");
         }}
-        isLoading={isLoading}
+        isLoading={isLoadingTable}
       />
       <Modal
         title={titleModal}
@@ -272,3 +290,4 @@ export const CarriersPage = () => {
     </>
   );
 };
+// 281
